refactor: Удален из интерфейса метод получения версии API
