+++ src/main/java/com/github/_1c_syntax/coverage41C/IDebugClient.java
@@ -4,8 +4,6 @@ import java.util.List;
 import java.util.UUID;
 
 public interface IDebugClient {
-    String getApiVersion() throws DebugClientException;
-
     void connect(String password) throws DebugClientException;
 
     void connectTargets(List<String> debugAreaNames, List<DebugTargetType> debugTargetTypes) throws DebugClientException;
